Add files via upload
All the files are hosted on dropbox now. all you need is this 6kb script. It'll have you connect your device then download and install Snapchat. The reason for this is that Snapchat is an xapk but these aren't handled by adb tools. Without the tool you'll have to extract the apks manually and copy all the location paths to run the adb install-multiple command. I learned a bit more about powershell while working on this and was a fun little project.